Completed Step 3 Added packages and moved files to their appropriate packages.
