+++ M04/Assignment/Android_Sprint_Networking_Challenge/app/src/main/res/layout/activity_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DetailsActivity">
+    tools:context=".View.DetailsActivity">
 
     <TextView
         android:id="@+id/tv_pokemon_name"
