+++ M04/Assignment/Android_Sprint_Networking_Challenge/app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_margin="16dp"
-    tools:context=".MainActivity">
+    tools:context=".View.MainActivity">
 
     <EditText
         android:id="@+id/et_pokemon_name_or_id"
